2021/10/27/Building a Kanban Board App in FastAPI and React/backend/requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
asyncpg==0.24.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
